Thx
This is all the files that I have in Computer Graphics. I push it on Github for I 'm unwilling to see so many people  fall into despair especially the schoolchildren in the same school. However,I'm so sorry that the code is not unabridged,as I'm not allowed to complete it by my own ability.
Wish to help you!